added clone-deep-helpers.js; used native prototype error handling over Object.prototype.toString.call when applicable; no longer using value.constructor to acquire constructor functions
