+++ clone-deep.js
@@ -1,103 +1,4 @@
-/**
- * Contains the tag for various types.
- * @type {Object<string, string>}
- */
-export const Tag = Object.freeze({
-    ARGUMENTS: "[object Arguments]",
-    ARRAY: "[object Array]",
-    BIGINT: "[object BigInt]",
-    BOOLEAN: "[object Boolean]",
-    DATE: "[object Date]",
-    ERROR: "[object Error]",
-    FUNCTION: "[object Function]",
-    MAP: "[object Map]",
-    NUMBER: "[object Number]",
-    OBJECT: "[object Object]",
-    PROMISE: "[object Promise]",
-    REGEXP: "[object RegExp]",
-    SET: "[object Set]",
-    STRING: "[object String]",
-    SYMBOL: "[object Symbol]",
-    WEAKMAP: "[object WeakMap]",
-    WEAKSET: "[object WeakSet]",
-    ARRAYBUFFER: "[object ArrayBuffer]",
-    DATAVIEW: "[object DataView]",
-    FLOAT32: "[object Float32Array]",
-    FLOAT64: "[object Float64Array]",
-    INT8: "[object Int8Array]",
-    INT16: "[object Int16Array]",
-    INT32: "[object Int32Array]",
-    UINT8: "[object Uint8Array]",
-    UINT8CLAMPED: "[object Uint8ClampedArray]",
-    UINT16: "[object Uint16Array]",
-    UINT32: "[object Uint32Array]",
-    BIGINT64: "[object BigInt64Array]",
-    BIGUINT64: "[object BigUint64Array]"
-});
-
-/**
- * All prototypes of supported types.
- */
-export const supportedPrototypes = Object.freeze([
-    Array.prototype,
-    BigInt.prototype,
-    Boolean.prototype,
-    Date.prototype,
-    Error.prototype,
-    Function.prototype,
-    Map.prototype,
-    Number.prototype,
-    Object.prototype,
-    Promise.prototype,
-    RegExp.prototype,
-    Set.prototype,
-    String.prototype,
-    Symbol.prototype,
-    ArrayBuffer.prototype,
-    DataView.prototype,
-    Float32Array.prototype,
-    Float64Array.prototype,
-    Int8Array.prototype,
-    Int16Array.prototype,
-    Int32Array.prototype,
-    Uint8Array.prototype,
-    Uint8ClampedArray.prototype,
-    Uint16Array.prototype,
-    Uint32Array.prototype,
-    BigInt64Array.prototype,
-    BigUint64Array.prototype
-]);
-
-/**
- * Some native prototypes have properties that cannot be accessed or reassigned.
- * All such properties are stored here.
- */
-export const forbiddenProps = Object.freeze({
-    [Tag.FUNCTION]: { 
-        prototype: Function.prototype,
-        properties: ["caller", "callee", "arguments"]
-    },
-    [Tag.MAP]: {
-        prototype: Map.prototype,
-        properties: ["size"]
-    },
-    [Tag.SET]: {
-        prototype: Set.prototype,
-        properties: ["size"]
-    },
-    [Tag.SYMBOL]: {
-        prototype: Symbol.prototype,
-        properties: ["description"]
-    },
-    [Tag.ARRAYBUFFER]: {
-        prototype: ArrayBuffer.prototype,
-        properties: ["byteLength", "maxByteLength", "resizable", "detached"]
-    },
-    [Tag.DATAVIEW]: {
-        prototype: DataView.prototype,
-        properties: ["buffer", "byteLength", "byteOffset"]
-    }
-});
+import { getTag, getConstructor, Tag, supportedPrototypes, forbiddenProps } from "./clone-deep-helpers.js";
 
 /**
  * Used to log warnings.
@@ -155,40 +56,6 @@ const Warning = {
  * @type {symbol}
  */ 
 const TOP_LEVEL = Symbol("TOP_LEVEL");
-
-/**
- * Gets a "tag", which is an string which identifies the type of a value.
- * `Object.prototype.toString` returns a string like `"[object <Type>]"`,  where 
- * type is the type of the object. We refer this return value as the **tag**. 
- * Normally, the tag is determined by what `this[Symbol.toStringTag]` is, but 
- * the JavaScript specification for `Object.prototype.toString` requires that 
- * many native JavaScript objects return a specific tag if the object does not 
- * have the `Symbol.toStringTag` property. Also, classes introduced after ES6 
- * typically have their own non-writable `Symbol.toStringTag` property. This 
- * makes `Object.prototype.toString.call` a stronger type-check that 
- * `instanceof`.
- * 
- * @example
- * ```
- * const date = new Date();
- * console.log(date instanceof Date);  // true
- * console.log(tagOf(date));  // "[object Date]"
- * 
- * const dateSubclass = Object.create(Date.prototype);
- * console.log(dateSubclass instance Date);  // true;
- * console.log(tagOf(dateSubClass));  // "[object Object]"
- * 
- * // This is not a perfect type check because we can do:
- * dateSubclass[Symbol.toStringTag] = "Date"
- * console.log(tagOf(dateSubClass));  // "[object Date]"
- * ```
- * 
- * @param {any} value The value to get the tag of.
- * @returns {string} tag A string indicating the value's type.
- */
-function tagOf(value) {
-    return Object.prototype.toString.call(value);
-}
 
 /**
  * Returns `true` if tag is that of a TypedArray subclass, `false` otherwise.
@@ -440,7 +307,7 @@ function cloneInternalNoRecursion(_value, customizer, log, doThrow) {
          * Identifies the type of the value.
          * @type {String}
          */
-        const tag = tagOf(value);
+        const tag = getTag(value);
 
         if (forbiddenProps[tag] !== undefined
             && forbiddenProps[tag].prototype === value)
@@ -504,7 +371,7 @@ function cloneInternalNoRecursion(_value, customizer, log, doThrow) {
 
             else if ([Tag.BOOLEAN, Tag.DATE].includes(tag)) {
                 /** @type {BooleanConstructor|DateConstructor} */
-                const BooleanOrDateConstructor = value.constructor;
+                const BooleanOrDateConstructor = getConstructor(value);
 
                 cloned = assign(new BooleanOrDateConstructor(Number(value)), 
                                 parentOrAssigner, 
@@ -513,7 +380,7 @@ function cloneInternalNoRecursion(_value, customizer, log, doThrow) {
             }
             else if ([Tag.NUMBER, Tag.STRING].includes(tag)) {
                 /** @type {NumberConstructor|StringConstructor} */
-                const NumberOrStringConstructor = value.constructor;
+                const NumberOrStringConstructor = getConstructor(value);
 
                 cloned = assign(new NumberOrStringConstructor(value), 
                                 parentOrAssigner, 
@@ -591,13 +458,10 @@ function cloneInternalNoRecursion(_value, customizer, log, doThrow) {
             
             else if (isTypedArray(tag)) {
                 /** @type {import("./private-types").TypedArrayConstructor} */
-                const TypedArray = value.constructor;
-
-                /** @type {ArrayBufferConstructor} */
-                const ArrayBufferConstructor = value.buffer.constructor;
+                const TypedArray = getConstructor(value);
 
                 // copy data over to clone
-                const buffer = new ArrayBufferConstructor(
+                const buffer = new ArrayBuffer(
                     value.buffer.byteLength);
                 new Uint8Array(buffer).set(new Uint8Array(value.buffer));
                 
@@ -616,7 +480,7 @@ function cloneInternalNoRecursion(_value, customizer, log, doThrow) {
                 const originalMap = value;
 
                 /** @type {MapConstructor} */
-                const MapConstructor = value.constructor;
+                const MapConstructor = getConstructor(value);
 
                 const cloneMap = new MapConstructor;
 
@@ -640,7 +504,7 @@ function cloneInternalNoRecursion(_value, customizer, log, doThrow) {
                 const originalSet = value;
 
                 /** @type {SetConstructor} */
-                const SetConstructor = value.constructor;
+                const SetConstructor = getConstructor(value);
 
                 const cloneSet = new SetConstructor;
